% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSexpressionSet.R
\name{getProbesetsFromStats}
\alias{getProbesetsFromStats}
\title{getProbesetsFromStats returns a list of probesets (the rownames from the data matrix) for a restriction of a list of stat comparisons}
\usage{
getProbesetsFromStats(x, v = 1e-04, pos = "padj", mode = "less",
  Comparisons = NULL)
}
\arguments{
\item{v}{The cutoff value}

\item{pos}{The column in the stats tables to base the selection on}

\item{mode}{one of 'less', 'more', 'onlyless' or 'equals' default 'less' ( <= )}

\item{Comparisons}{A list of comparisons to check (all if left out)}
}
\value{
a list of probesets that shows an adjusted p value below 1e-4
}
\description{
getProbesetsFromStats returns a list of probesets (the rownames from the data matrix) for a restriction of a list of stat comparisons
}
\examples{
probes <- getProbesetsFromStats ( x, v=1e-4, pos="adj.P.Val" ) 
}

