% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSexpressionSet.R, R/SingleCellsNGS.R
\docType{methods}
\name{normalize}
\alias{normalize}
\alias{normalize,NGSexpressionSet-method}
\title{description of function normalize}
\usage{
normalize(object, ..., reads = 600, name = "normalized")

normalize(object, ..., reads = 600, name = "normalized")
}
\arguments{
\item{reads}{the required read depth}

\item{name}{the name of the new object}

\item{x}{The NGSexpressionSet}

\item{readCounts}{The number of reads from each bam file or another value you want to normalize the data to}

\item{to_gene_length}{FALSE whether or not to normalize the data to gene length}

\item{geneLengthCol}{the column in the annotation data.frame to (in addition) normalize the genes to (e.g. trancript length)}

\item{x}{The SingleCellsNGS object}
}
\value{
the normalized data set (original data stored in NGS$raw

the normalized data set (original data stored in slot 'raw'
}
\description{
normalize the expression data (sample wise)

normalize the expression data by subsampling as described in PMID 24531970
}

