% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellsNGS.R
\docType{data}
\name{old_LTHSC}
\alias{old_LTHSC}
\title{Read counts for the expression data described in PMID:26430063}
\format{SingleCellsNGS}
\usage{
old_LTHSC
}
\description{
The data was re-mapped against mouse mm10 using HISAT

and quantified using the R subreads package and then restricted the old_LTHSC single cells only.
}
\keyword{datasets}

