% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSexpressionSet.R
\docType{methods}
\name{preprocess}
\alias{preprocess}
\alias{preprocess,NGSexpressionSet-method}
\title{description of function preprocess}
\usage{
preprocess(x)
}
\arguments{
\item{x}{the NGSexpressionSet}

\item{percentile}{how many of the top genes to use (0.05)}

\item{tf}{further subselect the top genes using these genes}

\item{x}{the NGSexpressionSet}
}
\value{
a list of values - please read the source code of this function
}
\description{
calculates how many reads are consumed by the top genes performes a set of preprocess

function used by DEseq Do not use
}

