% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSexpressionSet.R
\name{preprocess}
\alias{preprocess}
\title{performes a set of preprocess function used by DEseq}
\usage{
preprocess(x)
}
\arguments{
\item{x}{the NGSexpressionSet
Do not use}
}
\description{
performes a set of preprocess function used by DEseq
}

