% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellsNGS.R
\docType{methods}
\name{z.score}
\alias{z.score}
\alias{z.score,SingleCellsNGS-method}
\title{description of function z.score}
\usage{
\S4method{z.score}{SingleCellsNGS}(m)
}
\arguments{
\item{x}{the SingleCellsNGS object}
}
\description{
Use most of'StefansExpressionSet'functionallity with minor changes to NGS data (like normalizing)

This package is mainly meant to plot the data - all important quantification or gene

annotation is performed using DEseq functionallity. Read a set of bam files and perform

the quantification (better do that without using this function!)
}

