% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGSexpressionSet.R
\name{z.score}
\alias{z.score}
\title{zscore the ngs dataset}
\usage{
z.score(m)
}
\arguments{
\item{m}{the NGSexpressionSet}
}
\value{
The z scored NGSexpressionSet
}
\description{
zscore the ngs dataset
}

